requirements.txt
fastapi
uvicorn
psycopg2-binary
google-generativeai
pytesseract
pdf2image
pillow
python-dotenv
google-genai
python-multipart